backend/requirements.txt
Django==5.0.6
drf-yasg==1.21.7
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-dateutil==2.9.0
PyYAML==6.0.1